+++ src/Orchestra/Foundation/AdminMenuHandler.php
@@ -44,7 +44,7 @@ class AdminMenuHandler
      * Construct a new handler.
      *
      * @param  \Orchestra\Foundation\Application   $app
-     * @param  \Orchestra\Resources\Environment    $resouces
+     * @param  \Orchestra\Resources\Environment    $resources
      * @param  \Illuminate\Translation\Translator  $translator
      */
     public function __construct(Application $app, Resources $resources, Translator $translator)