+++ docs/changes.md
@@ -10,6 +10,7 @@ title: Foundation Change Log
 * Open hyperlink to extension author's URL on a new window/tab.
 * Allow all type-hinting to resolve to available service locator in Orchestra Platform.
 * Improve DI on `Orchestra\Foundation\AdminMenuHandler`.
+* Eagerly attach `orchestra/memory` during installation.
 
 ### v2.1.1 {#v2-1-1}