+++ docs/changes.md
@@ -5,13 +5,14 @@ title: Foundation Change Log
 
 ## Version 2.1 {#v2-1}
 
-### v2.1.2@dev {#v2-1-2}
+### v2.1.2 {#v2-1-2}
 
 * Open hyperlink to extension author's URL on a new window/tab.
 * Allow all type-hinting to resolve to available service locator in Orchestra Platform.
 * Improve DI on `Orchestra\Foundation\AdminMenuHandler`.
 * Eagerly attach `orchestra/memory` during installation.
 * Multiple refactor.
+* Handle `orchestra/memory` on `orchestra.mail` service locator from `orchestra/foundation`.
 
 ### v2.1.1 {#v2-1-1}